Fixed a small compilation error related to fmt.

+++ src/main.cpp
@@ -99,7 +99,7 @@ auto main(int p_argc, char** p_argv) -> int {
             stderr,
             fmt::fg(fmt::color::red),
             "[FATAL ERROR]: Could not allocate memory for a buffer: {}. Vulkan error {}",
-            exception.what, exception.error_code
+            exception.what, static_cast<int>(exception.error_code)
         );
     }
 }
